+++ packages/ui/dnd/src/hooks/useDndBlock.ts
@@ -1,24 +1,55 @@
 import { useState } from 'react';
 import { getEmptyImage } from 'react-dnd-html5-backend';
 import { useEditorRef } from '@udecode/plate-core';
-import { useDragBlock } from './useDragBlock';
-import { useDropBlockOnEditor } from './useDropBlockOnEditor';
+import { DropLineDirection } from '../types';
+import { useDragNode, UseDragNodeOptions } from './useDragNode';
+import { useDropNode, UseDropNodeOptions } from './useDropNode';
 
-export const useDndBlock = ({
-  id,
-  blockRef,
-  removePreview,
-}: {
+export interface UseDndNodeOptions {
   id: string;
   blockRef: any;
   removePreview?: boolean;
+}
+
+export const useDndNode = ({
+  id,
+  type,
+  blockRef,
+  removePreview,
+}: {
+  /**
+   *
+   */
+  id: string;
+
+  /**
+   * The type of item being dragged.
+   */
+  type: UseDragNodeOptions['type'];
+
+  /**
+   * The reference to the block being dragged.
+   */
+  blockRef: UseDropNodeOptions['blockRef'];
+
+  /**
+   * Whether to remove the preview.
+   */
+  removePreview?: boolean;
+
+  drag?: UseDragNodeOptions;
+  drop?: UseDropNodeOptions;
 }) => {
   const editor = useEditorRef();
 
-  const [dropLine, setDropLine] = useState<'' | 'top' | 'bottom'>('');
+  const [dropLine, setDropLine] = useState<DropLineDirection>('');
 
-  const [{ isDragging }, dragRef, preview] = useDragBlock(editor, id);
-  const [{ isOver }, drop] = useDropBlockOnEditor(editor, {
+  const [{ isDragging }, dragRef, preview] = useDragNode(editor, {
+    item: { id },
+    type,
+  });
+  const [{ isOver }, drop] = useDropNode(editor, {
+    accept: type,
     id,
     blockRef,
     dropLine,
@@ -43,3 +74,5 @@ export const useDndBlock = ({
     dragRef,
   };
 };
+
+export const useDndBlock = (options: UseDndNodeOptions) => useDndNode(options);
